# docs/06-api-and-integrations/artifacts/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "EndpointDTO": {
        "properties": {
          "end_line": {
            "title": "End Line",
            "type": "integer"
          },
          "file_path": {
            "title": "File Path",
            "type": "string"
          },
          "handler": {
            "title": "Handler",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Metadata",
            "type": "object"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "start_line": {
            "title": "Start Line",
            "type": "integer"
          }
        },
        "required": [
          "method",
          "path",
          "handler",
          "file_path",
          "start_line",
          "end_line"
        ],
        "title": "EndpointDTO",
        "type": "object"
      },
      "EndpointListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EndpointDTO"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "EndpointListResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImportantFileDTO": {
        "properties": {
          "file_path": {
            "title": "File Path",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "file_path",
          "reason"
        ],
        "title": "ImportantFileDTO",
        "type": "object"
      },
...
      "ModuleDTO": {
        "properties": {
          "file_count": {
            "default": 0,
            "title": "File Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "name",
          "summary"
        ],
        "title": "ModuleDTO",
        "type": "object"
      },
      "OverviewResponse": {
        "properties": {
          "detected_stack": {
            "items": {
              "type": "string"
            },
            "title": "Detected Stack",
            "type": "array"
          },
          "documentation_gaps": {
            "items": {
              "type": "string"
            },
            "title": "Documentation Gaps",
            "type": "array"
          },
          "endpoints": {
            "items": {
              "$ref": "#/components/schemas/EndpointDTO"
            },
            "title": "Endpoints",
            "type": "array"
          },
          "important_files": {
            "items": {
              "$ref": "#/components/schemas/ImportantFileDTO"
            },
            "title": "Important Files",
            "type": "array"
          },
          "modules": {
            "items": {
              "$ref": "#/components/schemas/ModuleDTO"
            },
            "title": "Modules",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "repository_id": {
            "title": "Repository Id",
            "type": "string"
          },
          "stats": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Stats",
            "type": "object"
          }
        },
        "required": [
          "repository_id",
          "name",
          "detected_stack",
          "important_files",
          "modules",
          "endpoints",
          "documentation_gaps",
          "stats"
        ],
        "title": "OverviewResponse",
        "type": "object"
      },
      "ReadingPathItemDTO": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "string"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "file_path": {
            "title": "File Path",
            "type": "string"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "signals": {
            "items": {
              "$ref": "#/components/schemas/ReadingPathSignalDTO"
            },
            "title": "Signals",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "rank",
          "file_path",
          "title",
          "reason",
          "confidence"
        ],
        "title": "ReadingPathItemDTO",
        "type": "object"
      },
      "ReadingPathResponse": {
        "properties": {
          "index_version": {
            "default": 0,
            "title": "Index Version",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReadingPathItemDTO"
            },
            "title": "Items",
            "type": "array"
          },
          "repository_id": {
            "title": "Repository Id",
            "type": "string"
          }
        },
        "required": [
          "repository_id",
          "items"
        ],
        "title": "ReadingPathResponse",
        "type": "object"
      },
      "ReadingPathSignalDTO": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "detail"
        ],
        "title": "ReadingPathSignalDTO",
        "type": "object"
      },
...
      "RepositoryDTO": {
        "properties": {
          "chunks": {
            "default": 0,
            "title": "Chunks",
            "type": "integer"
          },
          "current_index_version": {
            "default": 0,
            "title": "Current Index Version",
            "type": "integer"
          },
          "detected_stack": {
            "items": {
              "type": "string"
            },
            "title": "Detected Stack",
            "type": "array"
          },
          "endpoints": {
            "default": 0,
            "title": "Endpoints",
            "type": "integer"
          },
          "graph_nodes": {
            "default": 0,
            "title": "Graph Nodes",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "indexed_files": {
            "default": 0,
            "title": "Indexed Files",
            "type": "integer"
          },
          "last_indexed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Indexed At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "source_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Label"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          },
          "source_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Uri"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "symbols": {
            "default": 0,
            "title": "Symbols",
            "type": "integer"
          },
          "total_files": {
            "default": 0,
            "title": "Total Files",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "source_type",
          "status"
        ],
        "title": "RepositoryDTO",
        "type": "object"
      },
...
      "SymbolDTO": {
        "properties": {
          "end_line": {
            "title": "End Line",
            "type": "integer"
          },
          "file_path": {
            "title": "File Path",
            "type": "string"
          },
          "index_version": {
            "default": 0,
            "title": "Index Version",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "signature": {
            "default": "",
            "title": "Signature",
            "type": "string"
          },
          "start_line": {
            "title": "Start Line",
            "type": "integer"
          },
          "symbol_id": {
            "title": "Symbol Id",
            "type": "string"
          },
          "symbol_type": {
            "title": "Symbol Type",
            "type": "string"
          }
        },
        "required": [
          "symbol_id",
          "file_path",
          "symbol_type",
          "name",
          "start_line",
          "end_line"
        ],
        "title": "SymbolDTO",
        "type": "object"
      },
      "SymbolListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SymbolDTO"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "SymbolListResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AI Codebase Assistant",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/repositories": {
      "get": {
        "operationId": "list_repositories_api_v1_repositories_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RepositoryDTO"
                  },
                  "title": "Response List Repositories Api V1 Repositories Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Repositories",
        "tags": [
          "repositories"
        ]
      }
    },
...
    "/api/v1/repositories/{repository_id}/api/endpoints": {
      "get": {
        "operationId": "list_endpoints_api_v1_repositories__repository_id__api_endpoints_get",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "title": "Repository Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Endpoints",
        "tags": [
          "repositories"
        ]
      }
    },
...
    "/api/v1/repositories/{repository_id}/overview": {
      "get": {
        "operationId": "get_overview_api_v1_repositories__repository_id__overview_get",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "title": "Repository Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Overview",
        "tags": [
          "repositories"
        ]
      }
    },
    "/api/v1/repositories/{repository_id}/reading-path": {
      "get": {
        "operationId": "get_reading_path_api_v1_repositories__repository_id__reading_path_get",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "title": "Repository Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadingPathResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Reading Path",
        "tags": [
          "repositories"
        ]
      }
    },
...
    "/api/v1/repositories/{repository_id}/symbols": {
      "get": {
        "operationId": "list_symbols_api_v1_repositories__repository_id__symbols_get",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "title": "Repository Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "symbol_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol Type"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SymbolListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Symbols",
        "tags": [
          "repositories"
        ]
      }
    },
...
  }
}